Set up a bundling system, extract helper utils to 3rd party libs
